switch-screen: Downgrade FIXME to TODO, improve readability
Auto-configuration is difficult as we cannot easily guess which display
port the user is using, and also automatic resolution detection doesn't
seem to work well (as mentioned near the calls to xrandr).
So this isn't that important.
We should be okay to expect the user to do his bare minimum amount of
work of deciding what screen setup he wants anyway.